+++ src/main/java/kitchenpos/tablegroup/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.table.repository;
+package kitchenpos.tablegroup.repository;
 
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
